+++ src/middleware/auth.js
@@ -1,11 +1,11 @@
 import jwt from 'jsonwebtoken';
+let userData = '';
 
 const checkAuth = (req, res, next) => {
   try {
     const token = req.headers.authorization.split(' ')[1];
-
     const decoded = jwt.verify(token, process.env.JWT_KEY);
-    req.userData = decoded;
+    userData = decoded;
     next();
   } catch (err) {
     return res.status(401).json({
@@ -14,4 +14,6 @@ const checkAuth = (req, res, next) => {
   }
 };
 
-export default checkAuth;
+// const userId = userData.id;
+
+export { checkAuth, userData };
