+++ src/middleware/author-auth.js
@@ -1,0 +1,3 @@
+import { userData } from './auth';
+
+const userId = userData.id; // getting the user id from
